refactor: Custom header 재사용/ 불필요한 파일 삭제
