+++ src/components/InvestmentAccountList/InvestmentAccountList.tsx
@@ -6,14 +6,17 @@ import {
   useGetAccountQuery,
   usePrefetchAccountQuery,
 } from "@src/components/InvestmentAccountList/Account-query/InvestmentAccount.query";
-import InvestmentAccountTableHead from "./InvestmentAccountTableHead/InvestmentAccountTableHead";
 import InvestmentAccountItem from "./InvestmentAccountItem/InvestmentAccountItem";
 
 import Dropdown from "../common/Dropdown/Dropdown";
 import SearchInput from "../common/SearchInput/SearchInput";
-import PagenationButton from "./PagenationButton/PagenationButton";
+import PagenationButton from "../common/PagenationButton/PagenationButton";
 import { DROPDOWN_DATA } from "@src/constants/dropDownData";
 import Loader from "../common/Loader/Loader";
+import CustomTableHead from "../common/Table/CustomTableHead";
+import { ACCOUNT_TABLE_CELL_DATA } from "@src/constants/tableData";
+
+import * as S from "../UserList/UserList.style";
 
 const InvestmentAccountList = () => {
   const [accountQueryParams, setAccountQueryParams] = useAtom(
@@ -54,31 +57,33 @@ const InvestmentAccountList = () => {
 
   return (
     <>
-      <Container>
-        <SearchInput
-          onUpdateParams={setAccountQueryParams}
-          text="계좌명 검색"
-        />
-        {DROPDOWN_DATA.map(({ id, name, data }) => (
-          <Dropdown
-            accountQueryParams={accountQueryParams}
-            key={id}
-            setAccountQueryParams={setAccountQueryParams}
-            name={name}
-            data={data}
+      <S.Container>
+        <S.FilterContainer>
+          <SearchInput
+            onUpdateParams={setAccountQueryParams}
+            text="계좌명 검색"
           />
-        ))}
-        <Button
-          sx={{ margin: "8px" }}
-          variant="contained"
-          onClick={initDropDownFilters}
-        >
-          필터링 초기화
-        </Button>
-      </Container>
+          {DROPDOWN_DATA.map(({ id, name, data }) => (
+            <Dropdown
+              accountQueryParams={accountQueryParams}
+              key={id}
+              setAccountQueryParams={setAccountQueryParams}
+              name={name}
+              data={data}
+            />
+          ))}
+          <Button
+            sx={{ margin: "8px" }}
+            variant="contained"
+            onClick={initDropDownFilters}
+          >
+            필터링 초기화
+          </Button>
+        </S.FilterContainer>
+      </S.Container>
       <TableContainer component={Paper}>
         <Table sx={{ minWidth: 650 }} aria-label="simple table">
-          <InvestmentAccountTableHead pageType="InvestmentAccounts" />
+          <CustomTableHead data={ACCOUNT_TABLE_CELL_DATA} />
           <InvestmentAccountItem data={defaultAccountListData} />
         </Table>
       </TableContainer>
