+++ src/components/UserList/UserList.tsx
@@ -5,7 +5,7 @@ import { useAtom } from 'jotai';
 import { Table, TableContainer, Paper } from '@mui/material';
 
 import SearchInput from '../common/SearchInput/SearchInput';
-import PagenationButton from '../InvestmentAccountList/PagenationButton/PagenationButton';
+import PagenationButton from '../common/PagenationButton/PagenationButton';
 import {
   useDeleteUsers,
   useGetUserListQuery,
@@ -31,8 +31,8 @@ const UserList = () => {
   const { data, isLoading, isError } = useGetUserListQuery(userQueryParams);
   const isMaxPage = usePrefetchUserListQuery(userQueryParams).data?.data.length;
   const { mutate: deleteUser } = useDeleteUsers();
-  // usePrefetchAccountList(currentPage, maxPage);
 
+  
   const handleCheck = (userId: string) => {
     checked.includes(userId)
       ? setChecked(checked.filter(el => el !== userId))
