--- src/constants/UserTableData.ts
@@ -1,23 +1,0 @@
-export const USER_TABLE_DATA = [
-  { id: 1, text: '계좌수' },
-  { id: 2, text: '이메일 주소' },
-  { id: 3, text: '성별' },
-  { id: 4, text: '생년월일' },
-  { id: 5, text: '휴대폰 번호' },
-  { id: 6, text: '최근로그인' },
-  { id: 7, text: '수신 동의' },
-  { id: 8, text: '계좌활성화' },
-  { id: 9, text: '가입일' },
-];
-
-export const ACCOUNT_TABLE_HEAD = [
-  { id: 1, text: '고객명' },
-  { id: 2, text: '브로커명' },
-  { id: 3, text: '계좌번호' },
-  { id: 4, text: '계좌상태' },
-  { id: 5, text: '계좌명' },
-  { id: 6, text: '평가금액' },
-  { id: 7, text: '입금금액' },
-  { id: 8, text: '계좌활성화' },
-  { id: 9, text: '계좌개설일' },
-];
